Remove old IOTestEntity
