--- Assets/Scripts/Entities/IOEntities/IOButton.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7fd80606f2fbeee41b370da13739ec87
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
