--- Assets/Scripts/Entities/IOEntities/IOButton/IOButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 274d5d7d3e754ef41af1956809abae99
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
